KeyboardSimulator/requirements.txt
pynput
pillow
numpy
